# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "T-Loyal API",
        "title": "Crazy Peppers T-Loyal API",
        "contact": {},
        "version": "1.0"
    },
    "host": "prod-team-29-4254c2ee.REDACTED",
    "basePath": "/api/v1",
    "paths": {
...
        "/partners/{partnerID}/promos": {
            "get": {
                "description": "Returning list of promos of specific partner for client",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PromosForClient"
                ],
                "summary": "Promo endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Partner ID",
                        "name": "partnerID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization header must be set for access",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Promo"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
        "/partners/{partnerID}/promos/{promoID}": {
            "get": {
                "description": "Returning promo for client",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PromosForClient"
                ],
                "summary": "Promo endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Partner ID",
                        "name": "partnerID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Partner ID",
                        "name": "promoID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization header must be set for access",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Promo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
        "/partners/{partnerID}/promos/{promoID}/img": {
            "get": {
                "description": "Returns the image associated with a specific promo",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "PromosForClient"
                ],
                "summary": "Get promo image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Partner ID",
                        "name": "partnerID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Promo ID",
                        "name": "promoID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization header must be set for access",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image file data",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "error: Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "error: Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "error: Изображение не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/promos": {
            "get": {
                "description": "Get promo by partner",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PromosForPartners"
                ],
                "summary": "Promo endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization header must be set for access",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Promo"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            },
            "post": {
                "description": "Create promo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PromosForPartners"
                ],
                "summary": "Promo endpoint",
                "parameters": [
                    {
                        "description": "Promo",
                        "name": "promo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Promo"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization header must be set for access",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Promo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
...
        "/promos/{promoID}": {
            "get": {
                "description": "Get promo by partner",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PromosForPartners"
                ],
                "summary": "Promo endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Promo ID",
                        "name": "promoID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization header must be set for access",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Promo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete promo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PromosForPartners"
                ],
                "summary": "Promo endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Promo ID",
                        "name": "promoID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization header must be set for access",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
        "/promos/{promoID}/img/upload": {
            "post": {
                "description": "Allows partners to upload an image for a specific promo",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PromosForPartners"
                ],
                "summary": "Upload promo image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Promo ID",
                        "name": "promoID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Image file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization header must be set for access",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Файл успешно загружен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error: Файл обязателен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "error: Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/promos/{promoID}/{clientID}": {
            "get": {
                "description": "Get promo by partner",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PromosForPartners"
                ],
                "summary": "Promo endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Promo ID",
                        "name": "promoID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "clientID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization header must be set for access",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PromoCLientView"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Condition": {
            "type": "object",
            "properties": {
                "buy": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "COMMON",
                        "UNIQUE",
                        "BUNDLE"
                    ]
                }
            }
        },
...
        "models.Promo": {
            "type": "object",
            "required": [
                "condition",
                "image_id"
            ],
            "properties": {
                "condition": {
                    "$ref": "#/definitions/models.Condition"
                },
                "condition_string": {
                    "description": "✅ ИСПРАВЛЕНО ЗДЕСЬ!",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image_id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_approved": {
                    "type": "boolean"
                },
                "partner_id": {
                    "type": "string"
                },
                "promo_id": {
                    "type": "string"
                },
                "time_limit_end": {
                    "type": "integer"
                },
                "time_limit_start": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "usage_limit": {
                    "type": "integer"
                }
            }
        },
        "models.PromoCLientView": {
            "type": "object",
            "required": [
                "condition",
                "image_id"
            ],
            "properties": {
                "buy": {
                    "type": "integer"
                },
                "condition": {
                    "$ref": "#/definitions/models.Condition"
                },
                "condition_string": {
                    "description": "✅ ИСПРАВЛЕНО ЗДЕСЬ!",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "current": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "image_id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_approved": {
                    "type": "boolean"
                },
                "overall": {
                    "type": "integer"
                },
                "partner_id": {
                    "type": "string"
                },
                "promo_id": {
                    "type": "string"
                },
                "time_limit_end": {
                    "type": "integer"
                },
                "time_limit_start": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "usage_limit": {
                    "type": "integer"
                }
            }
        },
...
        "utils.Err": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}